extract delivery details validation to common file refactor certificates delivery details controller to use common validation create certs copies controller change back link on delivery details template to be dynamic based on the service user is in extract delivery details template as now a common file
