+++ src/controllers/certified-copies/delivery.details.controller.ts
@@ -1,16 +1,74 @@
 import { NextFunction, Request, Response } from "express";
+import { validationResult } from "express-validator";
+import { getAccessToken, getUserId } from "../../session/helper";
+import { CERTIFIED_COPY_DELIVERY_DETAILS } from "../../model/template.paths";
 import { createLogger } from "ch-structured-logging";
 import { APPLICATION_NAME } from "../../config/config";
+import { Basket } from "ch-sdk-node/dist/services/order/basket/types";
+import { getBasket } from "../../client/api.client";
+import { deliveryDetailsValidationRules, validate } from "../../utils/delivery-details-validation";
+
+const FIRST_NAME_FIELD: string = "firstName";
+const LAST_NAME_FIELD: string = "lastName";
+const ADDRESS_LINE_ONE_FIELD: string = "addressLineOne";
+const ADDRESS_LINE_TWO_FIELD: string = "addressLineTwo";
+const ADDRESS_TOWN_FIELD: string = "addressTown";
+const ADDRESS_COUNTY_FIELD: string = "addressCounty";
+const ADDRESS_POSTCODE_FIELD: string = "addressPostcode";
+const ADDRESS_COUNTRY_FIELD: string = "addressCountry";
 
 const logger = createLogger(APPLICATION_NAME);
 
 export const render = async (req: Request, res: Response, next: NextFunction): Promise<void> => {
     try {
-        res.status(200).json({ value: true });
+        const userId = getUserId(req.session);
+        const accessToken: string = getAccessToken(req.session);
+        const basket: Basket = await getBasket(accessToken);
+        // logger.info(`Get certificate item, id=${certificateItem.id}, user_id=${userId}, company_number=${certificateItem.companyNumber}`);
+        return res.render(CERTIFIED_COPY_DELIVERY_DETAILS, {
+            firstName: basket.deliveryDetails?.forename,
+            lastName: basket.deliveryDetails?.surname,
+            addressLineOne: basket.deliveryDetails?.addressLine1,
+            addressLineTwo: basket.deliveryDetails?.addressLine2,
+            addressCountry: basket.deliveryDetails?.country,
+            addressTown: basket.deliveryDetails?.locality,
+            addressPoBox: basket.deliveryDetails?.poBox,
+            addressPostcode: basket.deliveryDetails?.postalCode,
+            addressCounty: basket.deliveryDetails?.region,
+            templateName: CERTIFIED_COPY_DELIVERY_DETAILS
+        });
     } catch (err) {
         logger.error(`${err}`);
         next(err);
     }
 };
 
-export default [render];
+const route = async (req: Request, res: Response, next: NextFunction) => {
+    const errors = validationResult(req);
+    const errorList = validate(errors);
+    const firstName: string = req.body[FIRST_NAME_FIELD];
+    const lastName: string = req.body[LAST_NAME_FIELD];
+    const addressLineOne: string = req.body[ADDRESS_LINE_ONE_FIELD];
+    const addressLineTwo: string = req.body[ADDRESS_LINE_TWO_FIELD];
+    const addressTown: string = req.body[ADDRESS_TOWN_FIELD];
+    const addressCounty: string = req.body[ADDRESS_COUNTY_FIELD];
+    const addressPostcode: string = req.body[ADDRESS_POSTCODE_FIELD];
+    const addressCountry: string = req.body[ADDRESS_COUNTRY_FIELD];
+
+    if (!errors.isEmpty()) {
+        return res.render(CERTIFIED_COPY_DELIVERY_DETAILS, {
+            ...errorList,
+            addressCountry,
+            addressCounty,
+            addressLineOne,
+            addressLineTwo,
+            addressPostcode,
+            addressTown,
+            firstName,
+            lastName,
+            templateName: (CERTIFIED_COPY_DELIVERY_DETAILS)
+        });
+    }
+};
+
+export default [...deliveryDetailsValidationRules, route];
