+++ src/model/template.paths.ts
@@ -7,10 +7,10 @@ export const COLLECTION: string = "collection";
 export const CERTIFICATE_INDEX: string = `${certRoot}/index`;
 export const CERTIFICATE_OPTIONS: string = `${certRoot}/options`;
 export const CERTIFICATE_CHECK_DETAILS: string = `${certRoot}/check-details`;
-export const CERTIFICATE_DELIVERY_DETAILS: string = `${certRoot}/delivery-details`;
+export const CERTIFICATE_DELIVERY_DETAILS: string = "delivery-details";
 export const CERTIFICATE_ORDER_DETAILS: string = `${certRoot}/order-details`;
 
 export const CERTIFIED_COPY_INDEX: string = `${certifiedCopyRoot}/index`;
 export const CERTIFIED_COPY_CHECK_DETAILS: string = `${certifiedCopyRoot}/check-details`;
-export const CERTIFIED_COPY_DELIVERY_DETAILS: string = `${certifiedCopyRoot}/delivery-details`;
+export const CERTIFIED_COPY_DELIVERY_DETAILS: string = "delivery-details";
 export const CERTIFIED_COPY_ORDER_DETAILS: string = `${certifiedCopyRoot}/order-details`;
