+++ src/routers/certified-copies/routers.ts
@@ -1,16 +1,15 @@
 import { Router, Response, NextFunction, Request } from "express";
 
 import { ROOT_CERTIFIED_COPY, CERTIFIED_COPY_DELIVERY_DETAILS, CERTIFIED_COPY_CHECK_DETAILS } from "../../model/page.urls";
-
+import deliveryDetailsController, { render as renderCertifiedCopies } from "../../controllers/certified-copies/delivery.details.controller";
 import homeController from "../../controllers/certified-copies/home.controller";
 import checkDetailsController from "../../controllers/certified-copies/check.details.controller";
-import deliveryDetailsController from "../../controllers/certified-copies/delivery.details.controller";
 
 const router: Router = Router();
 
 router.get(ROOT_CERTIFIED_COPY, homeController);
 
-router.get(CERTIFIED_COPY_DELIVERY_DETAILS, deliveryDetailsController);
+router.get(CERTIFIED_COPY_DELIVERY_DETAILS, renderCertifiedCopies);
 router.post(CERTIFIED_COPY_DELIVERY_DETAILS, deliveryDetailsController);
 
 router.get(CERTIFIED_COPY_CHECK_DETAILS, checkDetailsController);
