+++ src/views/delivery-details.html
@@ -10,7 +10,7 @@
 {% block beforeContent %}
   {{ govukBackLink({
     text: "Back",
-    href: "certificate-options"
+    href: DELIVERY_DETAILS_BACK_LINK
   }) }}
 {% endblock %}
